improve setup.sh

+++ setup.sh
@@ -2,7 +2,41 @@
 
 mkdir -vp ~/.config
 
-# create symlinks
-ln -s ~/.dotfiles/.zshrc ~/.zshrc
-ln -s ~/.dotfiles/nvim/ ~/.config/nvim/
-mkdir -vp ~/.config/tmux/ && ln -s ~/.dotfiles/tmux.conf ~/.config/tmux/tmux.conf
+echo "✨ Checking dependencies..."
+
+if ! command -v luarocks &>/dev/null; then
+  echo "ℹ️ luarocks not found, installing..."
+  brew install luarocks
+else
+  echo "✅ luarocks found"
+fi
+
+if ! command -v rg &>/dev/null; then
+  echo "ℹ️ rg not found, installing..."
+  brew install ripgrep
+else
+  echo "✅ ripgrep found"
+fi
+
+echo "✨ Checking symlinks..."
+
+if [ ! -e "$HOME/.zshrc" ]; then
+  echo "ℹ️ ~/.zshrc not found, creating symlink..."
+  ln -s ~/.dotfiles/.zshrc ~/.zshrc
+else
+  echo "✅ ~/.zshrc found"
+fi
+
+if [ ! -e "$HOME/.config/nvim" ]; then
+  echo "ℹ️ ~/.config/nvim/ not found, creating symlink..."
+  ln -s ~/.dotfiles/nvim/ ~/.config/nvim/
+else
+  echo "✅ ~/.config/nvim/ found"
+fi
+
+if [ ! -e "$HOME/.config/tmux" ]; then
+  echo "ℹ️ ~/.config/tmux/ not found, creating symlink..."
+  mkdir -vp ~/.config/tmux/ && ln -s ~/.dotfiles/tmux.conf ~/.config/tmux/tmux.conf
+else
+  echo "✅ ~/.config/tmux/ found"
+fi
